- Added major changes in the database error handling.
- More GUI bug fixes and error handling.

+++ forms/heavenlybody/heavenlybodyoverview.cpp
@@ -38,8 +38,18 @@ void HeavenlyBodyOverview::doubleClicked(QModelIndex modelIndex)
 {
     if (heavenlyBodyModel->isEntitySelected())
     {
-        HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, true);
-        heavenlyBodyDetails->show();
+        try
+        {
+            HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, true);
+            heavenlyBodyDetails->show();
+        }
+        catch (const SqlQueryException &sqlQueryException)
+        {
+            QMessageBox::critical(this,
+                                  "Database SQL error",
+                                  QString("There was an error with an SQL statement!\n\nError:\n\n%1").arg(sqlQueryException.getSqlError()),
+                                  QMessageBox::Ok);
+        }
     }
 }
 
@@ -51,38 +61,61 @@ void HeavenlyBodyOverview::selectionChanged(const QItemSelection &selected, cons
 
 void HeavenlyBodyOverview::on_add_clicked()
 {
-    HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, false);
-    heavenlyBodyDetails->show();
+    try
+    {
+        HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, false);
+        heavenlyBodyDetails->show();
+    }
+    catch (const SqlQueryException &sqlQueryException)
+    {
+        QMessageBox::critical(this,
+                              "Database SQL error",
+                              QString("There was an error with an SQL statement!\n\nError:\n\n%1").arg(sqlQueryException.getSqlError()),
+                              QMessageBox::Ok);
+    }
 }
 
 void HeavenlyBodyOverview::on_edit_clicked()
 {
     if (heavenlyBodyModel->isEntitySelected())
     {
-        HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, true);
-        heavenlyBodyDetails->show();
+        try
+        {
+            HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, true);
+            heavenlyBodyDetails->show();
+        }
+        catch (const SqlQueryException &sqlQueryException)
+        {
+            QMessageBox::critical(this,
+                                  "Database SQL error",
+                                  QString("There was an error with an SQL statement!\n\nError:\n\n%1").arg(sqlQueryException.getSqlError()),
+                                  QMessageBox::Ok);
+        }
     }
 }
 
 void HeavenlyBodyOverview::on_deleteEntity_clicked()
 {
-    try
+    if (heavenlyBodyModel->getSelectedEntity())
     {
-        int result = QMessageBox::question(this,
-                                           "Delete a Heavenly Body",
-                                           QString("Would you like to delete the Heavenly Body '%1'?").arg(heavenlyBodyModel->getSelectedEntity()->getName()),
-                                           QMessageBox::Yes | QMessageBox::No);
-
-        if (result == QMessageBox::Yes)
+        try
         {
-            heavenlyBodyModel->deleteEntity();
+            int result = QMessageBox::question(this,
+                                               "Delete a Heavenly Body",
+                                               QString("Would you like to delete the Heavenly Body '%1'?").arg(heavenlyBodyModel->getSelectedEntity()->getName()),
+                                               QMessageBox::Yes | QMessageBox::No);
+
+            if (result == QMessageBox::Yes)
+            {
+                heavenlyBodyModel->deleteEntity();
+            }
         }
-    }
-    catch (const DeleteEntityFailedException &exception)
-    {
-        QMessageBox::critical(this,
-                              "Error while deleting",
-                              exception.getMessage() + "\n\nSQL-Error:\n" + exception.getSqlError(),
-                              QMessageBox::Ok);
+        catch (const DeleteEntityFailedException &exception)
+        {
+            QMessageBox::critical(this,
+                                  "Error while deleting",
+                                  exception.getMessage() + "\n\nSQL-Error:\n" + exception.getSqlError(),
+                                  QMessageBox::Ok);
+        }
     }
 }
