+++ forms/solarsystem/solarsystemoverview.cpp
@@ -38,8 +38,18 @@ void SolarSystemOverview::doubleClicked(QModelIndex modelIndex)
 {
     if (solarSystemModel->isEntitySelected())
     {
-        SolarSystemDetails *solarSystemDetails = new SolarSystemDetails(this, solarSystemModel, true);
-        solarSystemDetails->show();
+        try
+        {
+            SolarSystemDetails *solarSystemDetails = new SolarSystemDetails(this, solarSystemModel, true);
+            solarSystemDetails->show();
+        }
+        catch (const SqlQueryException &sqlQueryException)
+        {
+            QMessageBox::critical(this,
+                                  "Database SQL error",
+                                  QString("There was an error with an SQL statement!\n\nError:\n\n%1").arg(sqlQueryException.getSqlError()),
+                                  QMessageBox::Ok);
+        }
     }
 }
 
@@ -51,19 +61,50 @@ void SolarSystemOverview::selectionChanged(const QItemSelection &selected, const
 
 void SolarSystemOverview::on_add_clicked()
 {
-    SolarSystemDetails *solarSystemDetails = new SolarSystemDetails(this, solarSystemModel, false);
-    solarSystemDetails->show();
+    try
+    {
+        SolarSystemDetails *solarSystemDetails = new SolarSystemDetails(this, solarSystemModel, false);
+        solarSystemDetails->show();
+    }
+    catch (const SqlQueryException &sqlQueryException)
+    {
+        QMessageBox::critical(this,
+                              "Database SQL error",
+                              QString("There was an error with an SQL statement!\n\nError:\n\n%1").arg(sqlQueryException.getSqlError()),
+                              QMessageBox::Ok);
+    }
 }
 
 void SolarSystemOverview::on_edit_clicked()
 {
-    SolarSystemDetails *solarSystemDetails = new SolarSystemDetails(this, solarSystemModel, true);
-    solarSystemDetails->show();
+    try
+    {
+        SolarSystemDetails *solarSystemDetails = new SolarSystemDetails(this, solarSystemModel, true);
+        solarSystemDetails->show();
+    }
+    catch (const SqlQueryException &sqlQueryException)
+    {
+        QMessageBox::critical(this,
+                              "Database SQL error",
+                              QString("There was an error with an SQL statement!\n\nError:\n\n%1").arg(sqlQueryException.getSqlError()),
+                              QMessageBox::Ok);
+    }
 }
 
 void SolarSystemOverview::on_deleteEntity_clicked()
 {
-    solarSystemModel->deleteSolarSystem();
+    if (solarSystemModel->getCurrentSolarSystem())
+    {
+        int result = QMessageBox::question(this,
+                                           "Delete a Solar System",
+                                           QString("Would you like to delete the Solar System '%1'?").arg(solarSystemModel->getCurrentSolarSystem()->getName()),
+                                           QMessageBox::Yes | QMessageBox::No);
+
+        if (result == QMessageBox::Yes)
+        {
+            solarSystemModel->deleteSolarSystem();
+        }
+    }
 }
 
 void SolarSystemOverview::on_startSimulation_clicked()
